Updated particle js modes
