+++ src/components/about/about.css
@@ -28,6 +28,7 @@
     display: grid;
     grid-template-columns: repeat(3, 1fr);
     gap: 1.5rem;
+    z-index: 1;
 }
 
 .about__card {
