+++ src/components/contacts/contacts.css
@@ -9,6 +9,7 @@
     display: flex;
     flex-direction: column;
     gap: 1.2rem;
+    z-index: 1;
 }
 
 .contact__option {
