+++ src/components/experience/experience.css
@@ -12,6 +12,7 @@
     border-radius: 2rem;
     border: 1px solid transparent;
     transition: var(--transition);
+    z-index: 1;
 }
 
 .experience__container > div:hover {
