+++ src/components/portfolio/portfolio.css
@@ -1,6 +1,6 @@
 .portfolio__container {
     display: grid;
-    grid-template-columns: repeat(3, 1fr);
+    grid-template-columns: repeat(2, 1fr);
     gap: 1rem;
 }
 
@@ -10,6 +10,7 @@
     border-radius: 15px;
     border: 1px solid transparent;
     transition: var(--transition);
+    z-index: 1;
 }
 
 .portfolio__item:hover {
